requirements.txt
python-dotenv
langchain
openai
wikipedia
streamlit